requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
pydantic
langchain
langchain-google-genai
PyPDF2